Fixed ValueChanged signals from doulbe to int

+++ apps/vaporgui/FlowSubtabs.cpp
@@ -104,17 +104,17 @@ void FlowSeedingSubtab::_createSeedingSection(QWidget *parent) {
     _xSeedSliderEdit->SetIntType(true);
     _xSeedSliderEdit->SetRange(MIN_AXIS_SEEDS, MAX_AXIS_SEEDS);
     _griddedSeedsFrame->addWidget(new VLineItem("X axis seeds", _xSeedSliderEdit));
-    connect(_xSeedSliderEdit, SIGNAL(ValueChanged(double)), this, SLOT(_rakeNumOfSeedsChanged()));
+    connect(_xSeedSliderEdit, SIGNAL(ValueChanged(int)), this, SLOT(_rakeNumOfSeedsChanged()));
     _ySeedSliderEdit = new VSliderEdit();
     _ySeedSliderEdit->SetIntType(true);
     _ySeedSliderEdit->SetRange(MIN_AXIS_SEEDS, MAX_AXIS_SEEDS);
     _griddedSeedsFrame->addWidget(new VLineItem("Y axis seeds", _ySeedSliderEdit));
-    connect(_ySeedSliderEdit, SIGNAL(ValueChanged(double)), this, SLOT(_rakeNumOfSeedsChanged()));
+    connect(_ySeedSliderEdit, SIGNAL(ValueChanged(int)), this, SLOT(_rakeNumOfSeedsChanged()));
     _zSeedSliderEdit = new VSliderEdit();
     _zSeedSliderEdit->SetIntType(true);
     _zSeedSliderEdit->SetRange(MIN_AXIS_SEEDS, MAX_AXIS_SEEDS);
     _griddedSeedsFrame->addWidget(new VLineItem("Z axis seeds", _zSeedSliderEdit));
-    connect(_zSeedSliderEdit, SIGNAL(ValueChanged(double)), this, SLOT(_rakeNumOfSeedsChanged()));
+    connect(_zSeedSliderEdit, SIGNAL(ValueChanged(int)), this, SLOT(_rakeNumOfSeedsChanged()));
 
     // Rake selector
     _rakeWidget = new VGeometry2();
@@ -137,7 +137,7 @@ void FlowSeedingSubtab::_createSeedingSection(QWidget *parent) {
     _randomSeedsSliderEdit = new VSliderEdit(MIN_RANDOM_SEEDS, MAX_RANDOM_SEEDS);
     _randomSeedsSliderEdit->SetIntType(true);
     _randomSeedsFrame->addWidget(new VLineItem("Seed count", _randomSeedsSliderEdit));
-    connect(_randomSeedsSliderEdit, SIGNAL(ValueChanged(double)), this,
+    connect(_randomSeedsSliderEdit, SIGNAL(ValueChanged(int)), this,
             SLOT(_rakeNumOfSeedsChanged()));
 
     _biasWeightSliderEdit = new VSliderEdit(-1, 1, 0);
